feat: jdbc 를 위한 의존성 변경
