+++ src/main/java/wooteco/subway/controller/ExceptionAdvice.java
@@ -14,7 +14,7 @@ public class ExceptionAdvice {
     }
 
     @ExceptionHandler(Exception.class)
-    public ResponseEntity<String> unexpectedError(Exception exception) {
+    public ResponseEntity<String> unexpectedError() {
         return ResponseEntity.badRequest().body("예기치 못한 에러가 발생했습니다.");
     }
 }
