+++ src/main/java/wooteco/subway/repository/dao/StationDao.java
@@ -1,42 +1,18 @@
 package wooteco.subway.repository.dao;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import wooteco.subway.repository.entity.StationEntity;
 
-public class StationDao {
+public interface StationDao {
 
-    private static Long seq = 0L;
-    private static List<StationEntity> store = new ArrayList<>();
+    StationEntity save(final StationEntity stationEntity);
 
-    public StationEntity save(final StationEntity stationEntity) {
-        final StationEntity saved = new StationEntity(++seq, stationEntity.getName());
-        store.add(saved);
-        return saved;
-    }
+    List<StationEntity> findAll();
 
-    public List<StationEntity> findAll() {
-        return new ArrayList<>(store);
-    }
+    Optional<StationEntity> findByName(final String name);
 
-    public static void removeAll() {
-        store.clear();
-    }
+    Optional<StationEntity> findById(final Long id);
 
-    public Optional<StationEntity> findByName(final String name) {
-        return store.stream()
-                .filter(stationEntity -> stationEntity.getName().equals(name))
-                .findAny();
-    }
-
-    public Optional<StationEntity> findById(final Long id) {
-        return store.stream()
-                .filter(stationEntity -> stationEntity.getId().equals(id))
-                .findAny();
-    }
-
-    public void deleteById(final Long id) {
-        findById(id).ifPresent(stationEntity -> store.remove(stationEntity));
-    }
+    void deleteById(final Long id);
 }
