+++ src/main/java/wooteco/subway/service/StationService.java
@@ -1,7 +1,6 @@
 package wooteco.subway.service;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import wooteco.subway.domain.Station;
