+++ src/main/resources/applicaiton.yml
@@ -1,0 +1,9 @@
+spring:
+  datasource: # datasource 설정
+    driver-class-name: org.h2.Driver
+    url: jdbc:h2:mem:testdb
+    username: sa
+    password:
+  h2:
+    console:
+      enabled: true # h2 console 설정
