+++ src/test/java/wooteco/subway/acceptance/StationAcceptanceTest.java
@@ -16,14 +16,14 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import wooteco.subway.dto.response.StationResponseDto;
-import wooteco.subway.repository.dao.StationDao;
+import wooteco.subway.mockDao.MockStationDao;
 
 @DisplayName("지하철역 관련 기능")
 public class StationAcceptanceTest extends AcceptanceTest {
 
     @BeforeEach
     void initStore() {
-        StationDao.removeAll();
+        MockStationDao.removeAll();
     }
 
     @DisplayName("지하철역을 생성한다.")
